Update trajectory_controller.h
set_thruts returns int

+++ phx_trajectory_planner/include/phx_trajectory_planner/trajectory_controller.h
@@ -97,7 +97,7 @@ class trajectory_controller
         void set_thrusts();
 
         void do_controlling(ros::Publisher);
-        double convert_thrust(double newton);
+        int convert_thrust(double newton);
 };
 
 #endif
